% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pbbo-package.R
\docType{package}
\name{pbbo-package}
\alias{pbbo-package}
\title{pbbo: Prior by Bayesian Optimisation}
\description{
A thin wrapper around mlrMBO for choosing prior hyperparameters based on an ideal prior predictive distribution.
}
\author{
\strong{Maintainer}: Andrew Manderson \email{andrew.manderson@mrc-bsu.cam.ac.uk} (\href{https://orcid.org/0000-0002-4946-9016}{ORCID})

}
\keyword{internal}
